add common functions
